removed warnings

+++ data structures/static rmq/max_value.cpp
@@ -6,7 +6,7 @@ struct rMq {
 		t.resize(32-__builtin_clz(n),vector<int>(n));
 		for(int i = 0; i < n; ++i)
 			t[0][i] = a[i];
-		for(int k = 1, p = 1; k < t.size(); ++k, p<<=1)
+		for(int k = 1, p = 1; k < (int)t.size(); ++k, p<<=1)
 			for(int i = 0; i < n; ++i)
 				t[k][i] = (i+p<n)?max(t[k-1][i],t[k-1][i+p]):t[k-1][i];
 	}
