+++ data structures/static rmq/min_value.cpp
@@ -6,7 +6,7 @@ struct rmq {
 		t.resize(32-__builtin_clz(n),vector<int>(n));
 		for(int i = 0; i < n; ++i)
 			t[0][i] = a[i];
-		for(int k = 1, p = 1; k < t.size(); ++k, p<<=1)
+		for(int k = 1, p = 1; (int)k < t.size(); ++k, p<<=1)
 			for(int i = 0; i < n; ++i)
 				t[k][i] = (i+p<n)?min(t[k-1][i],t[k-1][i+p]):t[k-1][i];
 	}
